Major refactoring for coarse graining
Atoms and Bonds no longer exist as explicit objects. Instead, Atoms are
now referred to more generally as Particles which are the leaves of a
hierarchy. However, under the hood, these Particles are in fact simply
an alias for Compound objects. Every Compound now has a `graph`
attribute which stores connectivity between Particles that Compound
contains.