Fix bug in _translate_to()
I implemented this incorrectly. It was just moving the coordinates so
that their geometric center was at the origin. My unit tests didn't
catch it since I was comparing with translating a compound to the
origin. I added a new unit test that translates to a random point so
that this error would be caught.