Add optional 'scale' arg to apply_to_compound
An optional 'scale' argument has been added to the pattern's
apply_to_compound function. When scale=True, the previous default
behavior is used, where the pattern's points are scaled to the
lengths of the host compound's boundingbox and are shifted to the
mins of the boundingbox. When scale=False, the pattern's points are
used as is.