Add setter and getter for charge
With this functionality charge can be set for Compounds at the
bottom of the containment hierarchy, but returns an error if the
user attempts to set the charge of Compounds higher up in the
hierarchy. For Compounds higher in the hierarchy, charge returns
the sum of the charges on the particles the Compound contains.