Support hoomd v3 (#871)
* Remove unneeded imports

HOOMD imports most submodules and packages by default.
This change allows `import mbuild` to work with hoomd v3
installed.

* Remove unused imports

This allows `hoomd_simulation.py` to import when HOOMD v3 is installed.

* Add create_hoomd3_forcefield

* Fix typos

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Mention both supported mixing rules in the error

* Fix pre-commit errors

* Document snapshot as `hoomd.Snapshot`

* Wrap particles into the box for hoomd v3

* Document type of `hoomd_forcefield`

* Add PPPM support

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Rename hoomd3_forcefield to hoomd_forcefield

Recognize that hoomd v4 this code may not be incompatible with v4
or future major versions.

* Fix a bug in to_snapshot with hoomd v3

The next line of code assumes that `box` is a numpy array. In HOOMD v3,
it is a list.

* Start preparing for separate hoomd v3/v2 tests

Separate out mbuild tests that do not strongly depend on the HOOMD version
into TestHOOMDAny. Place the v2 specific tests in TestHOOMDv2 and skip
these tests when the HOOMD version is not 2.

* Port create_simulation tests to create_forcefield

Copy the relevant tests for create_simulation and modify them to
test create_forcefield.

* Run pre-commit

* Fix doc build errors

Don't check hoomd.__version__ on import. The sphinx autodoc mock
doesn't seem to support this.

* Add hoomd_forcefield to the documentation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Attempt to fix failing hoomd v2 tests

* Attempt to fix readthedocs build warning

* Install the CPU build of HOOMD

This will reduce CI test time. The conda auto-selection is supposed
to do this automatically when there is no GPU on the system,
but the CI test rig for mbuild is currently installing the cuda toolkit.

* Fix failing hoomd2 test

* test hoomd-3 in Linux, python-3.7

* WIP- chnage displayName and  use mamba update for hoomdv3.

* Revert change to environment-dev.yml

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Umesh Timalsina <umesh.timalsina@vanderbilt.edu>